+++ game/player.py
@@ -1,18 +1,69 @@
 from game.tiles import *
+from collections import Counter
+from game.board import *
+class InvalidIndexException(Exception):
+    pass
 class Player:
-    def __init__(self, bag_tiles=None , nameplayer=None, player_id=None):
-        self.bag_tiles = bag_tiles
-        self.tiles = []
-        self.score = 0
-        #self.nameplayer = nameplayer
-        #self.player_id = player_id
+    def __init__(self,nameplayer="",id=0,score=0):
+        self.score = score
+        self.name = nameplayer
+        self.number = id
+        self.rack = []
+        self.tile_counter = Counter() 
     
-    def draw_tiles(self,bag,count):
-        drawn_tiles = bag.take(count)
-        self.tiles.extend(drawn_tiles)
+    def draw_tiles(self, tiles=[]):
+        self.rack.extend(tiles)
+
+    def set_nameplayer(self):
+        self.name = str(input("Ingrese su nombre: "))
+
+    def scoresum (self, score:int ):
+        self.score += score
+
+    def giveTiles(self, positions: list):
+        tiles = []
+        for i in range(len(positions)):
+            if positions[i] < len(self.rack):
+                tiles.append(self.rack.pop(positions[i]))
+                for j in range(len(positions)):
+                    if positions[i] < positions[j]:
+                        positions[j] -= 1
+            else:
+                raise InvalidIndexException("Invalid index:")
+        return tiles
     
-    def reset(self):
-        self.tiles = []  # Restablece las fichas del jugador.
+    def hasTiles(self, word: str):
+        # Comprueba si la palabra dada se puede formar utilizando las fichas de la estantería.
+        # Devuelve:True si la palabra se puede formar con las fichas de la estantería, False en caso contrario.
+        if len(word) == 0:
+            return True
+        while len(word) > 0:
+            usedTiles = Counter()
+            for letter in word:
+                if not self.has_letters_in_rack(letter, usedTiles):
+                    hasBlankTileUsed = False
+                    for i in range(len(self.rack)):
+                        if self.rack[i].letter == "?" and usedTiles[i] < 1:
+                            hasBlankTileUsed = True
+                            usedTiles[i] += 1
+                            break
+                    if not hasBlankTileUsed:
+                        return False
+        return True
+    
+    def has_letters_in_rack(self, word): 
+        tiles = [tiles.letter for tiles in self.rack]
+        for letter in word:
+            if letter in tiles:
+                tiles.remove(letter)
+            else:
+                return False
+        return True
+    def show_rack(self):
+        lectern = " | ".join(f"{tile.letter}:{tile.value}" for tile in self.rack)
+        indices = f"indx:" + " " * 3 + "  ".join(f'{i+1:^4}' for i, _ in enumerate(self.rack))
+        return f"Player ID: {self.number}\nScore: {self.score}\nAtril: {lectern} |\n{indices}"
+
         
     #def play_tile(self, tile):
     #    if tile in self.tiles:
@@ -22,46 +73,3 @@ class Player:
     
     #def display_hand(self):
     #    print("Player's hand: ", self.tiles.__str__())
-
-    def scoresum (self, score:int ):
-        self.score += score
-        #agregar suma del valor de palabras
-
-class Rack:
-    def __init__(self, bag):
-        self.rack = []
-        self.bag = bag
-        self.initialize()
-
-    def initialize(self):
-        #le da sus 7 fichas.
-        for i in range(7):
-            self.add_rack()
-    
-    def add_rack(self):
-        #toma las fichas de la bag y se las da al atril.
-        self.rack.append(self.bag.take(1))
-    
-    def get_rack(self):
-        self.get_letter = Tile()
-        #muestra el atril en fomato string.
-        return ", ".join(str(item.get_letter() for item in self.rack)) # join toma una secuencia de elementos y los concatena en una cadena, separándolos con el delimitador especificado (en este caso, la coma seguida de un espacio). convierte cada letra en una cadena utilizando str().
-
-    def get_rack_arr(self):
-       #Devuelve el rack como una matriz de tile
-        return self.rack
-
-    def remove_from_rack(self, tile):
-        #Quita una ficha de la estantería
-        if tile in self.rack:
-           self.rack.remove(tile)
-
-    def get_rack_length(self):
-        #Devuelve el número de fichas que quedan en la estantería.
-        return len(self.rack)
-    
-    def replenish_rack(self):
-        self.get_remaining_tiles = BagTiles()
-        #Añade fichas a la estantería después de un turno, de forma que la estantería tenga 7 fichas
-        while self.get_rack_length() < 7 and self.bag.get_remaining_tiles() > 0:
-            self.add_rack()